Provide show_help if libz is missing on backend node
If libz is present on the frontend, then messages sent to the
backend nodes will be compressed. If the messages are flagged
as having been compressed and we cannot decompress them (e.g.,
if libz isn't present on the backend), then output a help
message and abort. Otherwise, we'll try to unpack the compressed
message and just report an unpack error, which isn't very helpful.

Signed-off-by: Ralph Castain <rhc@pmix.org>